requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.21.1
atpublic==7.0.0
attrs==25.4.0
bigtree==1.0.3
billiard==4.2.4
blinker==1.9.0
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dask==2025.11.0
databricks-sdk==0.74.0
decorator==5.2.1
deprecation==2.1.0
dictdiffer==0.9.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.10
dvc==3.64.2
dvc-data==3.16.12
dvc-gs==3.0.2
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dynaconf==3.2.12
entrypoints==0.4
evidently==0.7.17
exceptiongroup==1.3.1
fairlearn==0.13.0
Faker==38.2.0
fastapi==0.124.0
feast==0.47.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
gcsfs==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-cloud-storage-control==1.8.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.5.5
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.6.1
librt==0.7.3
litestar==2.18.0
litestar-htmx==0.5.0
llvmlite==0.46.0
locket==1.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mlflow==3.7.0
mlflow-skinny==3.7.0
mlflow-tracing==3.7.0
mmh3==5.2.0
msgspec==0.20.0
multidict==6.7.0
multipart==1.3.0
mypy==1.19.0
mypy_extensions==1.1.0
narwhals==2.13.0
networkx==3.4.2
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
opentelemetry-api==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
partd==1.4.2
pathspec==0.12.1
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.1
plotly==5.24.1
polyfactory==3.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-click==1.9.4
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.5.9
semver==3.0.4
shap==0.49.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.45
sqlparse==0.5.4
sqltrie==0.11.2
starlette==0.50.0
statsmodels==0.14.6
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
toolz==1.1.0
tqdm==4.67.1
typeguard==4.4.4
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.1
uuid6==2025.0.1
uvicorn==0.38.0
uvicorn-worker==0.4.0
uvloop==0.22.1
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
